Better error handling

+++ powertools-sqs/src/main/java/software/amazon/lambda/powertools/sqs/internal/BatchContext.java
@@ -3,48 +3,73 @@ package software.amazon.lambda.powertools.sqs.internal;
 import java.util.ArrayList;
 import java.util.List;
 
-import com.amazonaws.services.lambda.runtime.events.SQSEvent;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 import software.amazon.awssdk.services.sqs.SqsClient;
 import software.amazon.awssdk.services.sqs.model.DeleteMessageBatchRequest;
 import software.amazon.awssdk.services.sqs.model.DeleteMessageBatchRequestEntry;
 import software.amazon.awssdk.services.sqs.model.GetQueueUrlRequest;
 import software.amazon.lambda.powertools.sqs.SQSBatchProcessingException;
 
+import static com.amazonaws.services.lambda.runtime.events.SQSEvent.SQSMessage;
+import static java.lang.String.format;
 import static java.util.stream.Collectors.toList;
 
 public final class BatchContext {
-    private List<SQSEvent.SQSMessage> success = new ArrayList<>();
-    private List<SQSEvent.SQSMessage> failures = new ArrayList<>();
-    private List<Exception> exceptions = new ArrayList<>();
-    private SqsClient client;
+    private static final Log LOG = LogFactory.getLog(BatchContext.class);
+
+    private final List<SQSMessage> success = new ArrayList<>();
+    private final List<SQSMessage> failures = new ArrayList<>();
+    private final List<Exception> exceptions = new ArrayList<>();
+    private final SqsClient client;
 
     public BatchContext(SqsClient client) {
         this.client = client;
     }
 
-    public void addSuccess(SQSEvent.SQSMessage event) {
+    public void addSuccess(SQSMessage event) {
         success.add(event);
     }
 
-    public void addFailure(SQSEvent.SQSMessage event, Exception e) {
+    public void addFailure(SQSMessage event, Exception e) {
         failures.add(event);
         exceptions.add(e);
+    }
+
+    public void processSuccessAndReset(final boolean suppressException) {
+        try {
+            if (hasFailures()) {
+
+                deleteSuccessMessage();
+
+                if (suppressException) {
+                    List<String> messageIds = failures.stream().map(SQSMessage::getMessageId).collect(toList());
+                    LOG.debug(format("[%s] records failed processing, but exceptions are suppressed. Failed messages %s", failures.size(), messageIds));
+                } else {
+                    throw new SQSBatchProcessingException(exceptions);
+                }
+            }
+        } finally {
+            reset();
+        }
     }
 
     private boolean hasFailures() {
         return !failures.isEmpty();
     }
 
-    private void cleanUpAndReset() {
-        DeleteMessageBatchRequest request = DeleteMessageBatchRequest.builder()
-                .queueUrl(url())
-                .entries(success.stream().map(m -> DeleteMessageBatchRequestEntry.builder()
-                        .id(m.getMessageId())
-                        .receiptHandle(m.getReceiptHandle())
-                        .build()).collect(toList()))
-                .build();
+    private void deleteSuccessMessage() {
+        if (!success.isEmpty()) {
+            DeleteMessageBatchRequest request = DeleteMessageBatchRequest.builder()
+                    .queueUrl(url())
+                    .entries(success.stream().map(m -> DeleteMessageBatchRequestEntry.builder()
+                            .id(m.getMessageId())
+                            .receiptHandle(m.getReceiptHandle())
+                            .build()).collect(toList()))
+                    .build();
 
-        client.deleteMessageBatch(request);
+            client.deleteMessageBatch(request);
+        }
     }
 
     private String url() {
@@ -57,21 +82,8 @@ public final class BatchContext {
     }
 
     private void reset() {
-        success = new ArrayList<>();
-        failures = new ArrayList<>();
-        exceptions = new ArrayList<>();
-    }
-
-    public void processSuccessAndReset(final boolean suppressException) {
-        if (hasFailures() && !suppressException) {
-            SQSBatchProcessingException exception = new SQSBatchProcessingException(exceptions);
-            cleanUpAndReset();
-            throw exception;
-        } else if (hasFailures()) {
-            // LOG
-            cleanUpAndReset();
-        } else {
-            reset();
-        }
+        success.clear();
+        failures.clear();
+        exceptions.clear();
     }
 }
